feat: Implement robust code formatting and enhance visualization EL-1009
- Introduced `scripts/format_code.py` for a reliable and independent formatting workflow using Ruff, bypassing Claude Code hooks.
- Updated `Makefile` to integrate the new formatting script with `format`, `format-all`, and `format-recent` targets.
- Integrated `finplot` into backtesting examples (`backtest_sota_dec2024.py`, `backtest_sota_jan2025.py`, `backtest_sota_nov2024.py`) for enhanced data visualization, including detailed charts, indicators, and realistic trade markers.
- Refactored `SOTAMomentum` strategy by moving it to a dedicated `strategies/sota` directory and updated import paths across backtest examples.
- Removed legacy and unused `.claude` settings, `README.md`, `SETUP.md`, `SRC_CLEANUP_SUMMARY.md`, `WORKSPACE_OVERVIEW.md`, `run.sh`, and `test_finplot.py` to streamline the repository.
- Improved strategy logging in `enhanced_profitable_strategy_v2.py` and `sota_momentum.py` by replacing `console.print` with `self.log.info` for better integration with NautilusTrader's logging system.
- Corrected P&L calculation logic in backtest examples to use `account_report` and `generate_fills_report()` for accurate performance tracking.
- Refined type hints and minor formatting in `signal_filter.py` and `migrate_cache.py`.
- Added `strategies` directory to `pyrightconfig.json` for comprehensive type checking.

 Thursday 2025-07-17 11:58:08 Vancouver GMT-07:00

+++ nautilus_test/scripts/migrate_cache.py
@@ -71,7 +71,7 @@ def get_cache_size(directory: Path) -> int:
     return total_size
 
 
-def format_size(size_bytes: int) -> str:
+def format_size(size_bytes: float) -> str:
     """Format bytes as human-readable string."""
     for unit in ["B", "KB", "MB", "GB", "TB"]:
         if size_bytes < 1024.0:
